Move Datafusion engine to a package
