+++ datafusion/core/tests/sqllogictests/src/engines/mod.rs
@@ -1,0 +1,2 @@
+pub mod datafusion;
+
